separate caches into files
