+++ src/main/java/me/doubledutch/pikadb/Column.java
@@ -3,6 +3,7 @@ package me.doubledutch.pikadb;
 import java.util.*;
 import java.io.*;
 
+
 public class Column{
 	private PageFile pageFile;
 	public int rootId;
