+++ src/main/java/me/doubledutch/pikadb/NoopPageCache.java
@@ -1,0 +1,7 @@
+package me.doubledutch.pikadb;
+
+class NoopPageCache implements PageCache{
+	public boolean Set(int id) {
+		return false;
+	}
+}
